Move common declaration to header file
6-3
